backend/requirements.txt
djangorestframework>=3.12.4
django-cors-headers>=3.7.0
pusher>=3.0.0
djangorestframework-simplejwt>=4.7.1
Pillow>=8.2.0
mysqlclient>=2.0.3
django-allauth==0.45.0
django-storages==1.11.1
boto3==1.18.6
